+++ test/test_pelicula.py
@@ -1,7 +1,6 @@
-from unittest import mock
-
 import pytest
-from src.pelicula import Pelicula, es_valida, read_avg_note_from_page
+
+from src.pelicula import Pelicula, read_avg_note_from_page
 
 
 @pytest.fixture
@@ -124,27 +123,3 @@ def test_scrap_prop_aprobados(film: Pelicula):
 def test_scrap_prop_aprobados_without_note(film_without_note: Pelicula):
     film_without_note.get_prop_aprobados()
     assert film_without_note.prop_aprobados == 0
-
-
-@mock.patch.object(es_valida, "__kwdefaults__", {'SET_VALID_FILM': 1})
-def test_valid_film(film: Pelicula):
-    assert film.valid()
-
-
-@mock.patch.object(es_valida, "__kwdefaults__", {'SET_VALID_FILM': 1})
-def test_not_valid_film():
-
-    thriller = Pelicula.from_id(571895)
-    assert not thriller.valid()
-
-    luxo = Pelicula.from_id(231149)
-    assert not luxo.valid()
-
-    cabina = Pelicula.from_id(762673)
-    assert not cabina.valid()
-
-    simpsons = Pelicula.from_id(372160)
-    assert not simpsons.valid()
-
-    isabel = Pelicula.from_id(593820)
-    assert not isabel.valid()
